Moved loglevel to utils due to importcycle
